requirements.txt
Django==5.0.3
gunicorn==21.2.0
whitenoise==6.6.0
django-debug-toolbar==4.3.0
django-modeltranslation==0.18.11

psycopg==3.1.18

redis==5.0.3
celery==5.3.6
